Convert osgi.html to markdown
Before this change the OSGI Starter page was basically invalid HTML because the complete html header meta data was included include_relative
* i used a converter to convert it to convert html to markdown and did a few manual corrections

TODO we should  remove the old html content from osgi.html to avoid duplicate content (which google does not like). But i was not sure if the old html is still needed. At the moment I just added a redirect to the new page in case somebody finds it via google or bookmark

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>