implement the since: version frontmatter variable
this allows us to display a little badge on the top of the page telling which bndtools version the current page assumes. this badge links to the respective changelog page in github e.g.  since: 7.1.0 links to https://github.com/bndtools/bnd/wiki/Changes-in-7.1.0

also updated the resolution-view.png to show the new features / buttons

and also  moved some images to the existing images folder

Signed-off-by: Christoph Rueger <chrisrueger@gmail.com>